Moved multipolygon validation from caom2 to caom2utils
