+++ caom2utils/caom2utils/__init__.py
@@ -8,6 +8,7 @@ from .fits2caom2 import *  # noqa
 from .legacy import *  # noqa
 from .wcs_util import * # noqa
 from .wcsvalidator import * # noqa
+from .caomvalidator import * # noqa
 
 
 import logging
